+++ SmartMealFinderB/src/main/java/hu/project/smartmealfinderb/Service/DailyProgressService.java
@@ -8,7 +8,7 @@ public interface DailyProgressService {
 
     DailyProgress findTodayProgress(User userId);
 
-    void createTodayProgress(User user, DietPlan plan, float weight, int caloriesConsumed, int proteinConsumed, int carbsConsumed, int fatsConsumed, String comment);
+    void createTodayProgress(User user, DietPlan plan, double weight, double caloriesConsumed, double proteinConsumed, double carbsConsumed, double fatsConsumed, String comment);
 
-    void updateTodayProgress(DailyProgress existingProgress, float weight, int caloriesConsumed, int proteinConsumed, int carbsConsumed, int fatsConsumed, String comment);
+    void updateTodayProgress(DailyProgress existingProgress, double weight, double caloriesConsumed, double proteinConsumed, double carbsConsumed, double fatsConsumed, String comment);
 }
